Creacion e implementacion de los servicios, creacion de los metodos get para las respectivas tablas y creacion del metodo post para la tabla mercancia

+++ src/main/java/com/softnexos/back/repository/CargosRepository.java
@@ -3,5 +3,5 @@ package com.softnexos.back.repository;
 import com.softnexos.back.model.Cargos;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CargosRepository extends JpaRepository<Cargos, Long> {
+public interface CargosRepository extends JpaRepository<Cargos, Integer> {
 }
